testbot/RestoreVM: A script to restore a TestBot VM.
This allows restoring a VM from either a borg repository (potentially
remote), or a set of compressed tar files.
The VM can be renamed during restoration, in which case its uuid is
also modified to avoid conflicts.
RestoreVM automatically waits for the VM to not be in use before
restoring it, and minimises the time it takes to perform the final step.
Finally it automatically restarts libvirt to take the new VM into
account.
RestoreVM can also extract the files to a temporary directory so they
can be checked or modified before use.

Signed-off-by: Francois Gouget <fgouget@codeweavers.com>