testbot: Allow multiple VMs to share access to a hypervisor domain.
This makes it possible to define two or more VM instances that have the
same VirtURI and VirtDomain values and differ only in their IdleSnapshot
field. This can be used to perform tests on multiple configurations of a
given Windows version, such as testing different locales, etc.

However a hypervisor domain can only have one current snapshot. So the
TestBot can only use one of the VM instances that depend on this
hypervisor domain at any time.
Concretely:
- VM::Status does not necessarily matches the state of the hypervisor
  domain. So VM::Status==off does not mean that the TestBot is free to
  revert it. The status of the other VMs using the same hypervisor
  domain much be checked first.
- checkidle cannot assume that a VM is idle just because the domain is
  powered on. It should check that the current snapshot matches
  IdleSnapshot first.
- And more...

Signed-off-by: Francois Gouget <fgouget@codeweavers.com>