testbot: Remove LibvirtDomain::_UpdateStatus().
It was needed to resynchronize the VM status field during the Engine
initialization. However nowadays the Engine delegates this to
LibvirtTool which does not want the VM status to be changed behind this
back and thus bypasses _UpdateStatus().
The bypass is explicit in the case of PowerOff(), $NoStatus=1, and
implicit in the RevertToSnapshot() case: _UpdateStatus() does nothing if
get_state() == STATE_RUNNING (which is the case for our VMs) and
$VM->Status ne "off".

Signed-off-by: Francois Gouget <fgouget@codeweavers.com>
Signed-off-by: Alexandre Julliard <julliard@winehq.org>