testbot: Better deal with $MaxVMsWhenIdle > $MaxActiveVMs.
Make sure we limit the number of running VMs to $MaxActiveVMs even if we
have more idle VMs ready, such as when $MaxVMsWhenIdle > $MaxActiveVMs.
For instance this allows preparing two or three VMs for future jobs,
while still running the tasks one by one once there is a job to run.
Also prevent the premature powering off of the extra idle VMs that are
going to be used next (like we do for VMs that will be needed for the
current jobs next steps).

Signed-off-by: Francois Gouget <fgouget@codeweavers.com>
Signed-off-by: Alexandre Julliard <julliard@winehq.org>