FIXME testbot: Allow multiple VMs to share access to a hypervisor domain.
This makes it possible to define two or more VM instances that have the
same VirtURI and VirtDomain values and differ only in their IdleSnapshot
field. This can be used to perform tests on multiple configurations of a
given Windows version, such as testing different locales, etc.

However a hypervisor domain only has one current snapshot. So the
TestBot can only use one of the VM instances that depend on this
hypervisor domain at any time.

An important consequence is that $VM->Status does not necessarily
matches the state of the hypervisor domain. So $VM->Status == off does
not mean that the TestBot is free to revert it. The status of the other
VMs using the same hypervisor domain much be checked first. Thisis
handled by some extra _CanScheduleOnVM() calls.

Signed-off-by: Francois Gouget <fgouget@codeweavers.com>