testbot/web: Use the TestBot records to show its activity.
The main index page is nice to browse the jobs but does not allow
getting a synthetic picture of what the TestBot is doing at a given
time. Even the VM table at the bottom falls short as it only shows the
current state and does not even show what task a given VM is running.

Each row of the new Activity page shows what each VM was doing at that
time, with color coding, links to the relevant task for running VMs,
and a count of the runnable and queued tasks.
Going down the table shows past history, allowing one to see how the
TestBot got in the current state.
The VMs are also grouped by host which makes it easier to verify that
the scheduler is obeying its operational parameters like the limits on
simultaneous running tasks, reverting VMs, etc.
While the activity page shows no sensitive information for now access to
it is restricted to users with a TestBot account.

Signed-off-by: Francois Gouget <fgouget@codeweavers.com>