testbot: Also record the number of blocked tasks.
A task is blocked if it should run on a retired or deleted VM, or one
which is undergoing maintenance.
Previously the runnable and queued task counters would include blocked
tasks which was wrong since, although they are queued, blocked tasks
cannot be run until some external actor makes the VM usable again.

Signed-off-by: Francois Gouget <fgouget@codeweavers.com>