testbot: Keep a record of the status of the VMs.
Each row of the new Records table allows storing part of the details
of an event or of the state of the TestBot. These pieces of information
are put together into groups and timestamped through the RecordGroups
table. Together these tables allow rebuilding the activity of the
TestBot.
The first user of these new tables is the job scheduler which uses them
to store the new VM status if it changed. This will allow rebuilding
and visualizing the activity of the TestBot for monitoring or debugging.

Signed-off-by: Francois Gouget <fgouget@codeweavers.com>
Signed-off-by: Alexandre Julliard <julliard@winehq.org>