testbot: Fix the CompareRecordGroups() documentation and variable names.
The RecordGroup Id is an AUTO_INCREMENT field. These never wrap!
Also, while that field is defined as an INT(6), it is a 32 bit integer
anyway so we will not run out of ids for a long time. Generally
speaking, for integers the display width (6 here) is meaningless as far
as the TestBot is concerned.

Also use more descriptive variable names. $a and $b were a bit
misleading too since they have special meaning in the context of the
sort() function, meaning which they don't have here.

Signed-off-by: Francois Gouget <fgouget@codeweavers.com>
Signed-off-by: Alexandre Julliard <julliard@winehq.org>