testbot: Purge old jobs based on their Submitted timestamp.
Although it would only happen in case of a major TestBot malfunction,
a Job could end up in a state where it will never complete and thus
where its Ended field never gets set. So deleting old jobs based on
their Ended field could leave us with zombies.
This cannot happen with the Submitted timestamp since that field is not
nullable.
Also, a Job run time is normally very small compared to the expiration
time (under a couple of hours compared to several days) so this does not
make a significant difference in when a job actually gets deleted.

Signed-off-by: Francois Gouget <fgouget@codeweavers.com>
Signed-off-by: Alexandre Julliard <julliard@winehq.org>