testbot: Fix the activity start and end times.
Even if the last job completed hours ago, the period covered by the data
we have in the jobs table always goes all the way up to now. Also it
starts with when the oldest job on record was submitted, whether that
job completed successfully or not.
Similarly the data in our activity records always covers everything up
until the present.

Signed-off-by: Francois Gouget <fgouget@codeweavers.com>
Signed-off-by: Alexandre Julliard <julliard@winehq.org>