testbot: Fix the duration of some VM operations.
Sometimes a RecordGroup containing VM status information is followed by
a RecordGroup containing only VM result information. This typically
happens when a VM is running a task which ends in the 'boterror' state.
However this VM result RecordGroup does not mean the VM is no longer in
the 'running' state so it should not be used to set the previous
group's 'end' time.

Signed-off-by: Francois Gouget <fgouget@codeweavers.com>
Signed-off-by: Alexandre Julliard <julliard@winehq.org>