testbot: Allow restricting the statistics to the recent past.
Over long periods of time (days) and as long as the TestBot can keep
up, the job & task creation rate should be almost identical to the
job and task completion rate (modulo build errors for tasks). But for
short periods (under an hour) they can be significantly different.
Specifically the job and task completion rates are capped by the
TestBot's processing speed, while the job and task creation rates are
not.
This means jobs must be filtered based on their Submitted field when
counting new jobs, and on their Ended field when dealing with completed
jobs.
Rename the job and task counters to newjobs and newtasks to emphasize
that they count new jobs and tasks.

Signed-off-by: Francois Gouget <fgouget@codeweavers.com>
Signed-off-by: Alexandre Julliard <julliard@winehq.org>