testbot: Explicitly track new tasks on a per VM host basis.
The number of tasks that got run on a VM host (running.time.count) is
not a proxy for the number of tasks that got created for that VM host
since not all tasks get run, for instance if the build fails.
Furthermore, when computing rates over short periods, the rate at which
tasks get completed on a given VM host is not the same as the rate at
which they get created, just as it is for jobs.

Signed-off-by: Francois Gouget <fgouget@codeweavers.com>
Signed-off-by: Alexandre Julliard <julliard@winehq.org>