testbot: Add a Munin plugin to monitor the TestBot activity.
This will make it possible to compare current data to historic data for
crucial parameters like the number of jobs and tasks per hour, the job
completion time, each VM host's busy rate, etc.
It will also make it possible to see if the reverts are progressively
getting slower, or if on the contrary they got slower on a specific
date, which would point to some change on the host, etc.
Note that access to the new Munin web page requires an API key to limit
who can query these statistics.

Signed-off-by: Francois Gouget <fgouget@codeweavers.com>
Signed-off-by: Alexandre Julliard <julliard@winehq.org>