testbot/testagentd: Show a message when testagentd is restarted.
It can be useful to set up a VM to auto-login on Windows and have
TestAgentd start automatically on boot. However in such a configuration
it can be hard to detect when a test causes Windows to reboot: the
client reconnects automatically just like after a network glitch and the
VM desktop looks normal.
So with this patch TestAgentd can keep track of its restarts, lets the
client query that value, and brings up a dialog if an unexpected restart
occurs.
The start count can also be overriden to deal with cases where some
automated process causes a Windows reboot as part of the VM set up
process.

Signed-off-by: Francois Gouget <fgouget@codeweavers.com>
Signed-off-by: Alexandre Julliard <julliard@winehq.org>