testbot: Fix the main WineTestBot modules circular dependencies.
The main WineTestBot modules, those that implement access to the
database, are prone to having circular dependencies. This is for
instance the case between Jobs and Patches.  Failure to handle these
can force the use of fully qualified function names in hard to predict
cases.  So to avoid these issues, systematically declare the exports
in a BEGIN block so they happen early enough.

Signed-off-by: Francois Gouget <fgouget@codeweavers.com>
Signed-off-by: Alexandre Julliard <julliard@winehq.org>