testbot: Move the job scheduler to WineTestBot::Engine::Scheduler.
The job scheduler is specific to the TestBot engine so ScheduleJobs()
should not be in the Jobs package which is meant to be used by any
process that deals with jobs.
The scheduler could even be part of Engine.pl but since it's pretty
self-contained it makes more sense to put it in its own package, hence
WineTestBot::Engine:Scheduler.

Signed-off-by: Francois Gouget <fgouget@codeweavers.com>
Signed-off-by: Alexandre Julliard <julliard@winehq.org>