testbot: Detect and kill stuck task scripts.
The tasks themselves have a timeout which the corresponding scripts
enforce. However the scripts themselves may get stuck, typically due
to network problems. When that happens they can end up blocking the
whole TestBot. So make sure the TestBot engine itself can detect stuck
scripts and take corrective action.
Note that the detection is not very timely but will happen at the
latest in the SafetyNet() function. This means there will be at most a
10 minutes delay.

Signed-off-by: Francois Gouget <fgouget@codeweavers.com>
Signed-off-by: Alexandre Julliard <julliard@winehq.org>