testbot: Fix handling of test unit additions/removals in patch series.
The TestBot used to only analyse the last patch in the series to
determine whether it impacts the tests or not. However patches to
resource files and helper dlls require having an up-to-date list of the
test units, and that list may be impacted by previous patches in the
series.

Signed-off-by: Francois Gouget <fgouget@codeweavers.com>
Signed-off-by: Alexandre Julliard <julliard@winehq.org>