testbot: Improve Job file storage.
So far files needed by a Step were stored in that Step's directory.
This means the files produced by a Step are stored in the directory of
the Step that needs them. Since multiple Steps may need the same file
this can result in duplication (and also in retrieving the same file
multiple times). This duplication is also needed for staging files if
they are needed by more than one Step.
This patch changes file storage so that files produced by a Step are
stored in that Step's directory. Conversely files needed by a Step are
found in the PreviousNo Step's directory, or in the Job's directory if
PreviousNo is not set. This avoids duplication both when retrieving
files from a VM and for staged files.
To simplify the code Step::GetFullFileName() converts Step::FileName
to an absolute path pointing to the Step's file. To ensure backward
compatibility with existing Jobs it supports both the old and the new
schemes.

Signed-off-by: Francois Gouget <fgouget@codeweavers.com>
Signed-off-by: Alexandre Julliard <julliard@winehq.org>