testbot: Make it possible to get an unfiltered clone of a collection.
Once a filter has been applied to a collection it cannot be removed.
This is particularly a problem on Detailref collections (such as
$Job->Steps) as the code that has set the filter may be anywhere.
So Clone() makes it possible to get an unfiltered copy of the
collection without modifying the original collection's filter, and
thus without impacting the code using it.

Signed-off-by: Francois Gouget <fgouget@codeweavers.com>
Signed-off-by: Alexandre Julliard <julliard@winehq.org>